Rename db handler file
